Update ScheduleFragment
Fix layout

Fill data

Implement on item click event

+++ app/src/main/java/ua/zp/rozklad/app/ui/MainActivity.java
@@ -329,7 +329,7 @@ public class MainActivity extends ActionBarActivity
     }
 
     @Override
-    public void onScheduleItemClicked(int scheduleItemId) {
-
+    public void onScheduleItemClicked(long scheduleItemId) {
+        // TODO: Start ScheduleItemActivity
     }
 }
